+++ chain/client/src/stateless_validation/validate.rs
@@ -283,14 +283,12 @@ fn validate_chunk_endorsement_signature(
     epoch_manager: &dyn EpochManagerAdapter,
     endorsement: &ChunkEndorsement,
 ) -> Result<(), Error> {
-    if epoch_manager.should_validate_signatures() {
-        let validator = epoch_manager.get_validator_by_account_id(
-            &endorsement.chunk_production_key().epoch_id,
-            &endorsement.account_id(),
-        )?;
-        if !endorsement.verify(validator.public_key()) {
-            return Err(Error::InvalidChunkEndorsement);
-        }
+    let validator = epoch_manager.get_validator_by_account_id(
+        &endorsement.chunk_production_key().epoch_id,
+        &endorsement.account_id(),
+    )?;
+    if !endorsement.verify(validator.public_key()) {
+        return Err(Error::InvalidChunkEndorsement);
     }
     Ok(())
 }
@@ -299,14 +297,12 @@ fn validate_witness_contract_code_request_signature(
     epoch_manager: &dyn EpochManagerAdapter,
     request: &ContractCodeRequest,
 ) -> Result<(), Error> {
-    if epoch_manager.should_validate_signatures() {
-        let validator = epoch_manager.get_validator_by_account_id(
-            &request.chunk_production_key().epoch_id,
-            &request.requester(),
-        )?;
-        if !request.verify_signature(validator.public_key()) {
-            return Err(Error::Other("Invalid witness contract code request signature".to_owned()));
-        }
+    let validator = epoch_manager.get_validator_by_account_id(
+        &request.chunk_production_key().epoch_id,
+        &request.requester(),
+    )?;
+    if !request.verify_signature(validator.public_key()) {
+        return Err(Error::Other("Invalid witness contract code request signature".to_owned()));
     }
     Ok(())
 }
@@ -315,12 +311,9 @@ fn validate_witness_contract_accesses_signature(
     epoch_manager: &dyn EpochManagerAdapter,
     accesses: &ChunkContractAccesses,
 ) -> Result<(), Error> {
-    if epoch_manager.should_validate_signatures() {
-        let chunk_producer =
-            epoch_manager.get_chunk_producer_info(accesses.chunk_production_key())?;
-        if !accesses.verify_signature(chunk_producer.public_key()) {
-            return Err(Error::Other("Invalid witness contract accesses signature".to_owned()));
-        }
+    let chunk_producer = epoch_manager.get_chunk_producer_info(accesses.chunk_production_key())?;
+    if !accesses.verify_signature(chunk_producer.public_key()) {
+        return Err(Error::Other("Invalid witness contract accesses signature".to_owned()));
     }
     Ok(())
 }