+++ chain/epoch-manager/src/validate.rs
@@ -30,10 +30,6 @@ pub fn validate_optimistic_block_relevant(
         return Err(Error::InvalidBlockHeight(block.height()));
     }
 
-    if !epoch_manager_adapter.should_validate_signatures() {
-        return Ok(true);
-    }
-
     // A heuristic to check signature of the block even if prev block hash is not saved yet
     let epoch_ids =
         match epoch_manager_adapter.get_epoch_id_from_prev_block(&block.prev_block_hash()) {