+++ core/primitives/src/transaction.rs
@@ -248,8 +248,7 @@ impl ValidatedTransaction {
         }
     }
 
-    /// This method should only be used for test purposes.  This is because
-    /// kv_runtime is not designed to do proper signature verification.
+    /// This method should only be used for test purposes.
     pub fn new_for_test(signed_tx: SignedTransaction) -> Self {
         Self(signed_tx)
     }